LayoutTests/platform/blackberry/editing/text-iterator/findString-markers-expected.txt

PASS successfullyParsed is true

TEST COMPLETE
PASS internals.markerCountForNode(e.firstChild, "textmatch") is 0
PASS internals.markerCountForNode(e.firstChild, "textmatch") is 0
PASS internals.markerCountForNode(e.firstChild, "textmatch") is 0
PASS internals.markerCountForNode(e.firstChild, "textmatch") is 1
PASS "d" is "d"
PASS internals.markerCountForNode(e.firstChild, "textmatch") is 3
PASS "e" is "e"
PASS "e" is "e"
PASS "e" is "e"
PASS internals.markerCountForNode(e.firstChild, "textmatch") is 1
PASS "f" is "f"
PASS internals.markerCountForNode(e.firstChild, "textmatch") is 1
PASS "T" is "T"
PASS internals.markerCountForNode(e.firstChild, "textmatch") is 0
PASS internals.markerCountForNode(e.firstChild, "textmatch") is 0
PASS internals.markerCountForNode(e.firstChild, "textmatch") is 1
PASS "x" is "x"
PASS internals.markerCountForNode(e.firstChild, "textmatch") is 1
PASS "y" is "y"
PASS internals.markerCountForNode(e.firstChild, "textmatch") is 1
PASS "z" is "z"
PASS internals.markerCountForNode(e.firstChild, "textmatch") is 1
PASS "." is "."
PASS internals.markerCountForNode(e.firstChild, "textmatch") is 1
PASS "a" is "a"
PASS internals.markerCountForNode(e.firstChild, "textmatch") is 1
PASS "b" is "b"
PASS internals.markerCountForNode(e.firstChild, "textmatch") is 1
PASS "c" is "c"
PASS internals.markerCountForNode(e.firstChild, "textmatch") is 1
PASS "d" is "d"
PASS internals.markerCountForNode(e.firstChild, "textmatch") is 3
PASS "e" is "e"
PASS "e" is "e"
PASS "e" is "e"
PASS internals.markerCountForNode(e.firstChild, "textmatch") is 1
PASS "f" is "f"
PASS internals.markerCountForNode(e.firstChild, "textmatch") is 2
PASS "T" is "T"
PASS "t" is "t"
PASS internals.markerCountForNode(e.firstChild, "textmatch") is 2
PASS "u" is "u"
PASS "u" is "u"
PASS internals.markerCountForNode(e.firstChild, "textmatch") is 1
PASS "v" is "v"
PASS internals.markerCountForNode(e.firstChild, "textmatch") is 1
PASS "x" is "x"
PASS internals.markerCountForNode(e.firstChild, "textmatch") is 1
PASS "y" is "y"
PASS internals.markerCountForNode(e.firstChild, "textmatch") is 1
PASS "z" is "z"
PASS internals.markerCountForNode(e.firstChild, "textmatch") is 1
PASS "." is "."
The quick brown fox jumps over the lazy dog.
